+++ app/search/components/LawList/index.tsx
@@ -1,0 +1,3 @@
+import LawList from './LawList';
+
+export default LawList;
