+++ app/search/components/keyword/index.tsx
@@ -1,0 +1,3 @@
+import Keyword from './Keyword';
+
+export default Keyword;
