+++ app/statistics/components/Card/index.tsx
@@ -1,0 +1,3 @@
+import CardItem from './CardItem';
+
+export default CardItem;
