--- app/statistics/components/keyword/Keyword.tsx
@@ -1,9 +1,0 @@
-import Image from 'next/image';
-
-export default function Keyword() {
-  return (
-    <div className="flex justify-center">
-      <Image src="/mock/주요_키워드_이미지.svg" alt="국회 주요 키워드 이미지" width={349} height={271} />
-    </div>
-  );
-}
