+++ app/statistics/components/member/index.tsx
@@ -1,0 +1,4 @@
+import MemberItem from './MemberItem';
+import MemberList from './MemberList';
+
+export { MemberItem, MemberList };
