+++ app/statistics/components/party/index.tsx
@@ -1,0 +1,4 @@
+import Party from './Party';
+import PartyContainer from './PartyContainer';
+
+export { Party, PartyContainer };
